requirements.txt
torch>=1.9.0
torchvision>=0.10.0
gym>=0.21.0
numpy>=1.21.0
matplotlib>=3.3.0
Pillow>=8.0.0
scikit-image>=0.18.0
moviepy>=1.0.3
pympler>=0.9
py-desmume
